Moved key files to better locations
